Update index.html
I have added the html code for Rosa Parks and Maya Angelou. The dates I have chosen are 1955 and 1969 respectively, as this is when both women first gained recognition for their work/contributions. I have included the image links too, however please can you check that these are correct, as I was unable to access their links on my computer.